requirements.txt
Flask==3.0.3
numpy==2.1.2
Pillow==11.0.0
tensorflow==2.17.0

